pyproject.toml
[tool.poetry]
name = "gigaku"
version = "0.1.0"
description = "Migaku extension automation script"
package-mode = false

[tool.poetry.dependencies]
python = "^3.14"
selenium = "^4.40.0"
webdriver-manager = "^4.0.2"
pyobjc-framework-quartz = "^12.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
